conversai-backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
google-generativeai
edge-tts
mutagen
